move previous tests to separate file
