+++ tests/test_optimization/test_dmrg.py
@@ -1,83 +1,10 @@
 import numpy as np
 import scipy.sparse.linalg  # type: ignore
-from seemps.optimization.dmrg import QuadraticForm, dmrg
+from seemps.optimization.dmrg import dmrg
 from seemps.hamiltonians import ConstantTIHamiltonian, HeisenbergHamiltonian
-from seemps.cython import _contract_last_and_first
 from seemps.state import product_state, CanonicalMPS
-from seemps.operators import MPO
 from seemps.typing import DenseOperator
 from ..tools import SeeMPSTestCase
-
-
-class TestQuadraticForm(SeeMPSTestCase):
-    Sz: DenseOperator = np.diag([1, -1])
-    Sx: DenseOperator = np.array([[0, 1], [1, 0]])
-
-    def setUp(self) -> None:
-        return super().setUp()
-
-    def test_quadratic_form_checks_mpo_size(self):
-        mpo = MPO([np.ones((1, 2, 2, 1))] * 3)
-        mps = self.random_uniform_mps(2, 4)
-        with self.assertRaises(Exception):
-            QuadraticForm(mpo, mps)  # type: ignore
-
-    def test_quadratic_form_checks_mpo_dimensions(self):
-        mpo = MPO([np.ones((1, 2, 2, 1))] * 3)
-        mps = self.random_uniform_mps(3, 3)
-        with self.assertRaises(Exception):
-            QuadraticForm(mpo, mps)  # type: ignore
-
-    def test_quadratic_form_two_sites(self):
-        H = ConstantTIHamiltonian(size=2, interaction=np.kron(self.Sz, self.Sx))
-        Hmpo = H.to_mpo()
-        state = CanonicalMPS(self.random_uniform_mps(2, 2, D=2), center=0)
-
-        Q = QuadraticForm(Hmpo, state, start=0)
-        Hop = Q.two_site_Hamiltonian(0)
-        AB = _contract_last_and_first(state[0], state[1])
-        HopAB = Hop @ AB.reshape(-1)
-
-        self.assertEqual(HopAB.shape, (AB.size,))
-
-        expected = np.vdot(AB, HopAB)
-        exact_expected = Hmpo.expectation(state)
-        self.assertAlmostEqual(expected, exact_expected)  # type: ignore
-
-        HAB = H.to_matrix() @ AB.reshape(-1)
-        self.assertSimilar(HAB, HopAB)
-
-    def test_quadratic_form_three_sites_start_zero(self):
-        H = ConstantTIHamiltonian(size=3, interaction=np.kron(self.Sz, self.Sx))
-        Hmpo = H.to_mpo()
-        state = CanonicalMPS(self.random_uniform_mps(2, 3, D=2), center=0)
-
-        Q = QuadraticForm(Hmpo, state, start=0)
-        Hop = Q.two_site_Hamiltonian(0)
-        AB = _contract_last_and_first(state[0], state[1])
-        HopAB = Hop @ AB.reshape(-1)
-
-        self.assertEqual(HopAB.shape, (AB.size,))
-
-        expected = np.vdot(AB, HopAB)
-        exact_expected = Hmpo.expectation(state)
-        self.assertAlmostEqual(expected, exact_expected)  # type: ignore
-
-    def test_quadratic_form_three_sites_start_one(self):
-        H = ConstantTIHamiltonian(size=3, interaction=np.kron(self.Sz, self.Sx))
-        Hmpo = H.to_mpo()
-        state = CanonicalMPS(self.random_uniform_mps(2, 3, D=2), center=0)
-
-        Q = QuadraticForm(Hmpo, state, start=1)
-        Hop = Q.two_site_Hamiltonian(1)
-        AB = _contract_last_and_first(state[1], state[2])
-        HopAB = Hop @ AB.reshape(-1)
-
-        self.assertEqual(HopAB.shape, (AB.size,))
-
-        expected = np.vdot(AB, HopAB)
-        exact_expected = Hmpo.expectation(state)
-        self.assertAlmostEqual(expected, exact_expected)  # type: ignore
 
 
 class TestDMRG(SeeMPSTestCase):
